+++ com.unity.formats.fbx/Editor/Sources/Recorders/FbxRecorder/FbxRecorderSettingsEditor.cs
@@ -14,6 +14,25 @@ namespace UnityEditor.Formats.Fbx.Exporter
             EditorGUILayout.LabelField("Format", "FBX");
 
             FbxRecorderSettings settings = target as FbxRecorderSettings;
+
+            settings.ExportGeometry = EditorGUILayout.Toggle("Export Geometry", settings.ExportGeometry);
+        }
+
+        protected override void OnEncodingGui()
+        {
+            base.OnEncodingGui();
+
+            DrawSeparator();
+
+            EditorGUILayout.LabelField(new GUIContent(
+                "Transfer Animation",
+                "Transfer transform animation from source to destination. Animation on objects between source and destination will also be transferred to destination."
+            ));
+
+            FbxRecorderSettings settings = target as FbxRecorderSettings;
+
+            settings.TransferAnimationSource = EditorGUILayout.ObjectField("Source", settings.TransferAnimationSource, typeof(Transform), allowSceneObjects: true) as Transform;
+            settings.TransferAnimationDest = EditorGUILayout.ObjectField("Destination", settings.TransferAnimationDest, typeof(Transform), allowSceneObjects: true) as Transform;
         }
     }
 }